requirements.txt
transformers~=4.49.0.dev0
requests~=2.32.3
openai~=1.65.2
langchain~=0.3.20
python-dotenv~=1.0.1
pydantic~=2.10.6
uvicorn~=0.34.2
fastapi~=0.115.12